+++ framework/cmsunittests/PriorityQueueTest.cs
@@ -11,7 +11,7 @@ namespace cmsunittests
         [Test, Description("Call the Node<int>(priority, payload) constructor")]
         public void InitializeNodeInt()
         {
-            const float priority = 1.0f;
+            const double priority = 1.0;
             const int payload = 1;
 
             var node = new PriorityQueue<int>.Node<int>(priority, payload);
@@ -35,7 +35,7 @@ namespace cmsunittests
 
             var pq = new PriorityQueue<int>(size);
 
-            float showEntryReturn = pq.ShowEntry(-1);
+            double showEntryReturn = pq.ShowEntry(-1);
             //The above line should trigger an exception, if not, the test will fail.
         }
 
@@ -47,7 +47,7 @@ namespace cmsunittests
 
             var pq = new PriorityQueue<int>(size);
 
-            float showEntryReturn = pq.ShowEntry(size);
+            double showEntryReturn = pq.ShowEntry(size);
             //The above line should trigger an exception, if not, the test will fail.
         }
 
@@ -59,8 +59,8 @@ namespace cmsunittests
 
             var pq = new PriorityQueue<int>(size);
             
-            pq.Add(1.0f, size);
-            pq.UpdateIndex(1.4f, size+1);
+            pq.Add(1.0, size);
+            pq.UpdateIndex(1.4, size + 1);
             //The above line should trigger an exception, if not, the test will fail.
         }
 
@@ -71,13 +71,13 @@ namespace cmsunittests
 
             var pq = new PriorityQueue<int>(size);
 
-            pq.Add(1.0f, 0);
-            pq.Add(2.0f, 1);
+            pq.Add(1.0, 0);
+            pq.Add(2.0, 1);
             Assert.AreEqual(1, pq.LastIndex);
             pq.Clear();
             Assert.AreEqual(-1, pq.LastIndex);
-            pq.Add(1.0f, 2);
-            pq.Add(2.0f, 3);
+            pq.Add(1.0, 2);
+            pq.Add(2.0, 3);
             Assert.AreEqual(1, pq.LastIndex);
         }
 
@@ -90,15 +90,15 @@ namespace cmsunittests
             Assert.AreEqual(size, pq.Size);
             Assert.AreEqual(-1, pq.LastIndex);
 
-            pq.Add(1.0f, size);
+            pq.Add(1.0, size);
             Assert.AreEqual(0, pq.LastIndex);
             PriorityQueue<int>.Node<int> node = pq.First;
-            float priority;
+            double priority;
             int payload;
             pq.Top(out priority, out payload);
             Assert.AreEqual(node.Priority, priority);
             Assert.AreEqual(node.Payload, payload);
-            float showEntryReturn = pq.ShowEntry(size - 1);
+            double showEntryReturn = pq.ShowEntry(size - 1);
             Assert.AreEqual(priority, showEntryReturn);
 
             Console.WriteLine("Priority: " + priority);
@@ -107,7 +107,7 @@ namespace cmsunittests
             Console.WriteLine("Node.Payload: " + node.Payload);
             Console.WriteLine("ShowEntry.Priority: " + showEntryReturn);
 
-            pq.UpdateIndex(1.4f, size);
+            pq.UpdateIndex(1.4, size);
             node = pq.First;
             pq.Top(out priority, out payload);
             showEntryReturn = pq.ShowEntry(size-1);
@@ -131,13 +131,13 @@ namespace cmsunittests
             Assert.AreEqual(size, pq.Size);
             Assert.AreEqual(-1, pq.LastIndex);
 
-            float priority;
+            double priority;
             int payload;
 
             for (int i = 0; i < size; i++)
             {
                 pq.Add(i, i);
-                Assert.AreEqual((float)i, pq.ShowEntry(i));
+                Assert.AreEqual((double)i, pq.ShowEntry(i));
                 Assert.AreEqual(i, pq.LastIndex);
             }
 
@@ -147,15 +147,15 @@ namespace cmsunittests
             Assert.AreEqual(node.Payload, payload);
             Assert.AreEqual(priority, pq.ShowEntry(0));
 
-            pq.UpdateIndex(1.5f, 0);
+            pq.UpdateIndex(1.5, 0);
             node = pq.First;
             pq.Top(out priority, out payload);
             Assert.AreEqual(node.Priority, priority);
             Assert.AreEqual(node.Payload, payload);
             Assert.AreEqual(priority, pq.ShowEntry(0));
-            Assert.AreEqual(1.0f, priority);
+            Assert.AreEqual(1.0, priority);
             Assert.AreEqual(1, payload);
-            Assert.AreEqual(1.5f, pq.ShowEntry(1));
+            Assert.AreEqual(1.5, pq.ShowEntry(1));
         }
 
         [Test, Description("Priority > leftChildTau for index != 0")]
@@ -165,16 +165,16 @@ namespace cmsunittests
             var pq = new PriorityQueue<int>(size);
             Assert.AreEqual(size, pq.Size);
 
-            pq.Add(0.0f, 0);
-            pq.Add(5.0f, 1);
-            pq.Add(3.0f, 2);
-            pq.Add(4.0f, 3);
-            pq.Add(2.0f, 4);
+            pq.Add(0.0, 0);
+            pq.Add(5.0, 1);
+            pq.Add(3.0, 2);
+            pq.Add(4.0, 3);
+            pq.Add(2.0, 4);
             PrintNodes(pq);
 
-            pq.UpdateIndex(3.5f, 1);
-            pq.UpdateIndex(2.5f, 1);
-            pq.UpdateIndex(3.5f, 0);
+            pq.UpdateIndex(3.5, 1);
+            pq.UpdateIndex(2.5, 1);
+            pq.UpdateIndex(3.5, 0);
             PrintNodes(pq);
 
             PriorityQueue<int>.Node<int>[] nodes = pq.Nodes;
@@ -192,14 +192,14 @@ namespace cmsunittests
             var pq = new PriorityQueue<int>(size);
             Assert.AreEqual(size, pq.Size);
 
-            pq.Add(2.0f, 4);
-            pq.Add(2.5f, 1);
-            pq.Add(3.0f, 2);
-            pq.Add(6.0f, 0);
-            pq.Add(4.0f, 3);
+            pq.Add(2.0, 4);
+            pq.Add(2.5, 1);
+            pq.Add(3.0, 2);
+            pq.Add(6.0, 0);
+            pq.Add(4.0, 3);
             PrintNodes(pq);
 
-            pq.UpdateIndex(5.0f, 4);
+            pq.UpdateIndex(5.0, 4);
             PrintNodes(pq);
             PriorityQueue<int>.Node<int>[] nodes = pq.Nodes;
             Assert.AreEqual(2.5, nodes[0].Priority);
@@ -216,14 +216,14 @@ namespace cmsunittests
             var pq = new PriorityQueue<int>(size);
             Assert.AreEqual(size, pq.Size);
 
-            pq.Add(0.0f, 0);
-            pq.Add(2.0f, 4);
-            pq.Add(3.0f, 2);
-            pq.Add(2.5f, 1);
-            pq.Add(4.0f, 3);
+            pq.Add(0.0, 0);
+            pq.Add(2.0, 4);
+            pq.Add(3.0, 2);
+            pq.Add(2.5, 1);
+            pq.Add(4.0, 3);
             PrintNodes(pq);
 
-            pq.UpdateIndex(6.0f, 0);
+            pq.UpdateIndex(6.0, 0);
             PrintNodes(pq);
 
             PriorityQueue<int>.Node<int>[] nodes = pq.Nodes;
@@ -238,34 +238,34 @@ namespace cmsunittests
         public void PriorityQueueUpdateIndexTest12()
         {
             var pq = new PriorityQueue<int>(3);
-            pq.Add(1.0f, 1);
-            pq.Add(2.0f, 2);
-            pq.Add(3.0f, 3);
+            pq.Add(1.0, 1);
+            pq.Add(2.0, 2);
+            pq.Add(3.0, 3);
 
             // Update 1st node (payload 1) to be between 2nd and 3rd nodes
-            pq.UpdateIndex(2.5f, 1);
+            pq.UpdateIndex(2.5, 1);
 
             var nodes = pq.Nodes;
-            Assert.AreEqual(2.0f, nodes[0].Priority);
-            Assert.AreEqual(2.5f, nodes[1].Priority);
-            Assert.AreEqual(3.0f, nodes[2].Priority);
+            Assert.AreEqual(2.0, nodes[0].Priority);
+            Assert.AreEqual(2.5, nodes[1].Priority);
+            Assert.AreEqual(3.0, nodes[2].Priority);
         }
 
         [Test, Description("UpdateIndex, first node => last node")]
         public void PriorityQueueUpdateIndexTest13()
         {
             var pq = new PriorityQueue<int>(3);
-            pq.Add(1.0f, 1);
-            pq.Add(3.0f, 2);
-            pq.Add(2.0f, 3);
+            pq.Add(1.0, 1);
+            pq.Add(3.0, 2);
+            pq.Add(2.0, 3);
 
             // Update 1st node (payload 1) to be last
-            pq.UpdateIndex(4.0f, 1);
+            pq.UpdateIndex(4.0, 1);
 
             var nodes = pq.Nodes;
-            Assert.AreEqual(2.0f, nodes[0].Priority);
-            Assert.AreEqual(3.0f, nodes[1].Priority);
-            Assert.AreEqual(4.0f, nodes[2].Priority);
+            Assert.AreEqual(2.0, nodes[0].Priority);
+            Assert.AreEqual(3.0, nodes[1].Priority);
+            Assert.AreEqual(4.0, nodes[2].Priority);
         }
 
         private void PrintNodes<T>(PriorityQueue<T> pq)
